Clean up native proof audit findings
