+++ packages/runtime/src/native-hex.ts
@@ -1,0 +1,3 @@
+export function normalizeNativeHex(value: string): string {
+  return `0x${BigInt(value).toString(16)}`;
+}
